% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_miWQS_Functions.R
\name{check_constants}
\alias{check_constants}
\title{check_constants: K, B, T, n.burn all used this function}
\usage{
check_constants(K)
}
\arguments{
\item{K}{The number of imputed datasets, a natural number.}
}
\description{
check_constants: K, B, T, n.burn all used this function
}
