% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_wqs_helper.R
\name{print.wqs}
\alias{print.wqs}
\title{Prints the fitted WQS model along with the mean weights.}
\usage{
\method{print}{wqs}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An object of class WQS, from \code{\link{estimate.wqs}}.}

\item{digits}{minimal number of \emph{significant} digits, see
    \code{\link[base]{print.default}}.}

\item{...}{Further arguments passed from other methods to print(). Currently has no effect.}
}
\description{
Prints the fitted WQS model along with the mean weights. Adjusted from print.lm.
}
\examples{
# See estimate.wqs().

# As base package is always available, there is no need to ever import base
}
\seealso{
Other wqs: 
\code{\link{analyze.individually}()},
\code{\link{coef.wqs}()},
\code{\link{do.many.wqs}()},
\code{\link{estimate.wqs.formula}()},
\code{\link{estimate.wqs}()},
\code{\link{make.quantile.matrix}()},
\code{\link{plot.wqs}()}
}
\concept{wqs}
\keyword{wqs}
