% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqs.pool.test_data.R
\docType{data}
\name{wqs.pool.test}
\alias{wqs.pool.test}
\title{Combining WQS Regression Estimates}
\format{
An array of 16 x 2 x 3, with \itemize{
\item 16 parameters as the rows (The 14 weights, intercept, and WQS estimate of a WQS model),
\item 2 refers to the parameters of mean and standard deviation
\item K=3 complete imputed datasets.
}
}
\usage{
data(wqs.pool.test)
}
\description{
\emph{wqs.pool.test} was produced to demonstrate \code{pool.mi}(). First, the univariate Bayesian imputation approach (using \code{impute.univariate.bayesian.mi}) imputed the \code{X.bdl} element of \code{simdata87} multiple times to form an imputed X array. Multiple WQS regressions were run on the imputed X array to produce an array of WQS parameter estimates, \emph{wqs.pool.test}.
}
\examples{
wqs.pool.test <- data(wqs.pool.test)
# stage_3_pool_mi_example
}
\seealso{
\code{pool.mi}
}
\keyword{datasets}
