% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqs.pool.test_data.R
\docType{data}
\name{wqs.pool.test}
\alias{wqs.pool.test}
\title{Combining WQS Regression Estimates}
\format{An array of 16 x 2 x 3, with \itemize{
\item 16 parameters as the rows (The 14 weights, Intercept, and WQS  of the wqs.model),
\item 2 refers to the parameters of mean and standard deviation
\item K=3 imputation draws.
}}
\usage{
data(wqs.pool.test)
}
\description{
\emph{wqs.pool.test} was produced to demonstrate \code{pool.mi}(). The \code{simdata87} was first imputed multiple times to form an imputed X array. This was an example from \code{impute.univariate.bayesian.mi}. Multiple WQS regressions were run on the imputed X array to produce an array of WQS parameter estimates, \emph{wqs.pool.test}.
}
\examples{
wqs.pool.test <- data(wqs.pool.test)
}
\seealso{
\code{pool.mi}
}
\keyword{datasets}
